Do not error stopping vtadmin web and api if pidd files do not exist.
For fixing below.
https://github.com/vitessio/vitess/issues/13432

+++ examples/common/scripts/vtadmin-down.sh
@@ -2,8 +2,16 @@
 
 source "$(dirname "${BASH_SOURCE[0]:-$0}")/../env.sh"
 
-echo "Stopping vtadmin-web..."
-kill -9 "$(cat "$VTDATAROOT/tmp/vtadmin-web.pid")"
+if test -e "$VTDATAROOT/tmp/vtadmin-web.pid"; then
+  echo "Stopping vtadmin-web..."
+  kill -9 "$(cat "$VTDATAROOT/tmp/vtadmin-web.pid")"
+else
+  echo "Skipping stopping vtadmin-web because no pid file."
+fi
 
-echo "Stopping vtadmin-api..."
-kill -9 "$(cat "$VTDATAROOT/tmp/vtadmin-api.pid")"
+if test -e "$VTDATAROOT/tmp/vtadmin-api.pid"; then
+  echo "Stopping vtadmin-api..."
+  kill -9 "$(cat "$VTDATAROOT/tmp/vtadmin-api.pid")"
+else
+  echo "Skipping stopping vtadmin-web because no pid file."
+fi
